fix(launchpad): Move frequently used functions from Launchpad to Common Utilities
- https://github.com/gnoswap-labs/gnoswap-interface/pull/542#discussion_r1816047933
- https://github.com/gnoswap-labs/gnoswap-interface/pull/542#discussion_r1816041355
- https://github.com/gnoswap-labs/gnoswap-interface/pull/542#discussion_r1816037203
