--- packages/web/src/utils/launchpad-get-claimable-time.ts
@@ -1,5 +1,0 @@
-export const getClaimableTime = (threshold?: number) => {
-  if (threshold == null) return;
-
-  return new Date(Date.now() + threshold);
-};
